refactor: move useState variable (sort) to Main/index.tsx
