+++ static/onlyoffice-docs-confluence-cloud-custom-ui/src/components/ContentTree/utils/sortUtils.ts
@@ -1,0 +1,37 @@
+/**
+ *
+ * (c) Copyright Ascensio System SIA 2025
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import { SortOrder } from "../../../types/types";
+
+export const adoptSortForTargetRequest = (sort: {
+  key: string;
+  order: SortOrder;
+}) => {
+  const validValues = new Map([
+    ["title", "title"],
+    ["lastmodified", "modified-date"],
+  ]);
+
+  const value = validValues.get(sort.key);
+
+  if (value) {
+    return `${sort.order === "ASC" ? "" : "-"}${value}`;
+  }
+
+  return "";
+};
