+++ static/onlyoffice-docs-confluence-cloud-custom-ui/src/types/types.ts
@@ -20,6 +20,11 @@ import { CONTENT_TYPES } from "../constants";
 
 export type ContentType = (typeof CONTENT_TYPES)[number];
 
+export enum SortOrder {
+  ASC = "ASC",
+  DESC = "DESC",
+}
+
 export interface AppContext {
   appId: string;
   environmentId: string;
